+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - feat: add `--configure` option (to be implemented)
 - feat: add `--token` option
+- test: use `its_block` in rspec
 
 ## [v2.1.6] - 2025-01-05